fix: premake configuration to include Core/Debug dir and fix function type declarations

+++ core/src/Utils/Time.cpp
@@ -6,7 +6,7 @@
 
 namespace Ephemeral
 {
-    float Time::GetTime()
+    double Time::GetTime()
     {
         return glfwGetTime();
     }
